// src/infra/api/model/updatePreferencesBody.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Backend Store Check
 * API for storage and management of products and users
 * OpenAPI spec version: 1.0.0
 */
import type { UpdatePreferencesBodyTheme } from './updatePreferencesBodyTheme';

export type UpdatePreferencesBody = {
  email_notifications?: boolean;
  product_updates?: boolean;
  weekly_report?: boolean;
  theme?: UpdatePreferencesBodyTheme;
};