// src/infra/api/model/updateProfileBody.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Backend Store Check
 * API for storage and management of products and users
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateProfileBodyRole } from './updateProfileBodyRole';

export type UpdateProfileBody = {
  /** @minLength 1 */
  name: string;
  role: UpdateProfileBodyRole;
  branch_id?: string;
};